Add com.snowplowanalytics.mobile/diagnostic_error/jsonschema/1-0-0
It's a specification for an event generated by mobile trackers when they report an internal error.
It can be used as diagnostic of the tracker and it would be helpful to spot in advance anomalies in the tracker behaviour.
The diagnostic_error events can be collected by the customer and they are activated enabling a `diagnostic` feature in the mobile tracker.